Add the possiblity of using external sdcard as application storage for user 0
Since Android 4.2 with multiuser it is not possible to use external SD card for
application data (such as downloaded content for games). This patch adds
the possibility to use external SD card as a storage for user ID 0.

Mount point of the external sdcard is supposed to be defined by SECONDARY_STORAGE
environmental variable (exported by init on supported devices) and property
"persist.sys.env.use_sec_storage" is used to check if the storage is
switched. The property is to be set in "Advanced Settings" (for now)
defined under device tree (this should be available to user ID 0 only!)

Change-Id: I73d4711eac53118c0234110f06070fc42a975c44